% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ssh.R
\name{ssh_session}
\alias{channel_read}
\alias{channel_read_stderr}
\alias{channel_write}
\alias{channel_write_stderr}
\alias{ssh_session}
\title{SSH}
\usage{
ssh_session(host, port = 22, user = me(), key = "~/.ssh/id_rsa",
  password = readline, verbose = TRUE)

channel_read(session)

channel_read_stderr(session)

channel_write(session, x)

channel_write_stderr(session, x)
}
\arguments{
\item{host}{string with hostname}

\item{port}{integer}

\item{user}{username to authenticate with}

\item{key}{(optional) private RSA key to authenticate}

\item{password}{string or callback function with password}

\item{verbose}{emit some more output}
}
\description{
Interact with secure shell.
}

