% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ssh.R
\name{ssh}
\alias{ssh}
\title{Secure Shell}
\usage{
ssh(host, port = 22, user = me(), key = "~/.ssh/id_rsa",
  password = readline, verbose = FALSE)
}
\arguments{
\item{host}{string with hostname}

\item{port}{integer}

\item{user}{username to authenticate with}

\item{key}{(optional) private RSA key to authenticate}

\item{password}{string or callback function with password}

\item{verbose}{emit some more output}
}
\description{
Start an interactive SSH session.
}

